+++ include/spanningTree/pgr_kruskal.hpp
@@ -32,6 +32,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include <boost/graph/kruskal_min_spanning_tree.hpp>
 #include <vector>
 #include "spanningTree/pgr_mst.hpp"
+#include "cpp_common/interruption.h"
 
 namespace pgrouting {
 namespace functions {
@@ -70,7 +71,8 @@ template <class G>
 void
 Pgr_kruskal<G>::generate_mst(const G &graph) {
     this->clear();
-
+    /* abort in case of an interruption occurs (e.g. the query is being cancelled) */
+    CHECK_FOR_INTERRUPTS();
     boost::kruskal_minimum_spanning_tree(
             graph.graph,
             std::inserter(this->m_spanning_tree.edges, this->m_spanning_tree.edges.begin()),