+++ src/common/basic_vertex.cpp
@@ -129,7 +129,7 @@ namespace pgrouting {
     }
 
     std::vector < Basic_vertex > extract_vertices(
-            std::vector < Basic_vertex > vertices,
+            const std::vector < Basic_vertex > &vertices,
             const pgr_edge_t *data_edges, int64_t count) {
         return extract_vertices(
                 vertices, std::vector < pgr_edge_t >(