+++ tools/release-scripts/check-release.sh
@@ -52,22 +52,23 @@ fi
 
 
 MAYOR=3
-MINOR=0
+MINOR=1
 MICRO=0
-RC="-rc1"
-PREV_REL=2.6.3
+RC=""
+PREV_REL=3.0.0
 PREV_RELS="
+    3.0.0 3.0.1
     2.6.3 2.6.2 2.6.1 2.6.0
+    "
+# These releases are not for update
+OLD_RELS="
     2.5.4 2.5.3 2.5.2 2.5.1 2.5.0
     2.4.2 2.4.1 2.4.0
     2.3.2 2.3.1 2.3.0
-    2.2.4 2.2.3 2.2.2 2.2.1 2.2.0"
-# These releases are not for update
-OLD_RELS="
+    2.2.4 2.2.3 2.2.2 2.2.1 2.2.0
     2.1.0
     2.0.1 2.0.0"
 DEBUG=$1
-BRANCH="release/$MAYOR.$MINOR"
 
 
 if [[ -z  "$DEBUG" ]]; then
@@ -117,9 +118,8 @@ if [[ -n $DEBUG ]]; then
     echo
 fi
 
-CURRENTNEWS=$(grep $MAYOR.$MINOR.$MICRO doc/src/release_notes.rst | grep ref)
-if [[ $? != 0 ]]; then
-    error_msg "Section $MAYOR.$MINOR.$MICRO in release_notes.rst file is missing"
+if ! grep $MAYOR.$MINOR.$MICRO doc/src/release_notes.rst > /dev/null; then
+    rror_msg "Section $MAYOR.$MINOR.$MICRO in release_notes.rst file is missing"
     exit 1
 else
     echo "- [x] release_notes.rst section $MAYOR.$MINOR.$MICRO exists"
@@ -128,8 +128,7 @@ fi
 
 for r in $PREV_RELS $OLD_RELS
 do
-    OLDNEWS=$(grep $r doc/src/release_notes.rst | grep ref)
-    if [[ $? != 0 ]]; then
+    if ! grep "${r}" doc/src/release_notes.rst > /dev/null; then
         error_msg "Section $r in release_notes.rst file is missing"
         exit 1
     else
@@ -209,7 +208,7 @@ echo
 test_file "$MAYOR.$MINOR.$MICRO"
 for r in $PREV_RELS
 do
-    test_file $r
+    test_file "$r"
 done
 
 
@@ -234,11 +233,10 @@ echo "\`\`\`"
 echo bash tools/testers/update-tester.sh
 echo "\`\`\`"
 
-for r in $PREV_RELS
+for r in ${PREV_RELS}
 do
-    bash tools/testers/update-tester.sh "$r" "$MAYOR.$MINOR.$MICRO"
-    if [[ $? != 0 ]]; then
-        echo "FATAL on the update-tester"
+    if ! bash tools/testers/update-tester.sh "$r"; then
+        echo "FATAL updating from $r"
         exit 1
     fi
 done
@@ -253,10 +251,6 @@ echo tools/testers/doc_queries_generator.pl -documentation
 echo git status
 echo "\`\`\`"
 tools/testers/doc_queries_generator.pl -documentation
-if [[ $? != 0 ]]; then
-    echo "FATAL errors found generating documentation result files"
-    exit 1
-fi
 
 if [[ -z  "$DEBUG" ]]; then
     git_no_change