+++ tools/testers/update-tester.sh
@@ -26,7 +26,7 @@
 
 set -e
 
-TWEAK="-dev"
+TWEAK=""
 PGPORT=5432
 #sorry this only works on vicky's computer
 # TODO make it more general
@@ -37,42 +37,35 @@ function info {
 
     # ----------------------
     #
-    echo **pgRouting version must be installed before calling**
+    echo "**pgRouting version must be installed before calling**"
     echo "$1"
-    echo "$2"
 
-    echo EXAMPLE USAGE
-    echo - Short execution
-    echo  bash tools/testers/update-tester.sh $1 $2
-    echo - For running pgtap tests:
-    echo  bash tools/testers/update-tester.sh $1 $2 long
+    echo "EXAMPLE USAGE"
+    echo "- Short execution"
+    echo  "bash tools/testers/update-tester.sh $1"
+    echo "- For running pgtap tests:"
+    echo  "bash tools/testers/update-tester.sh $1 long"
 
 }
 
 
 if [[ -z  "$1" ]]; then
-    echo missing version to upgrade
-    info $CURRENT 2.6.3
-    exit 1
-fi
-
-if [[ -z  "$2" ]]; then
-    echo missing current version
-    info 2.6.3 3.0.0
+    echo missing version example:
+    info 2.6.3
     exit 1
 fi
 
 FROM_PGR="$1"
-CURRENT="$2"
-LONG=$3
+CURRENT=$(grep -Po '(?<=project\(PGROUTING VERSION )[^;]+' CMakeLists.txt)
+LONG=$2
 
 dropdb --if-exists "$DB"
 
 
 cd build
-cmake -DPGROUTING_DEBUG=ON -DCMAKE_BUILD_TYPE=Debug .. > /tmp/foo.txt
-make -j 4 > /tmp/foo.txt
-sudo make install > /tmp/foo.txt
+cmake -DPGROUTING_DEBUG=ON -DCMAKE_BUILD_TYPE=Debug ..
+make -j 4
+sudo make install
 cd ..
 
 
@@ -105,7 +98,7 @@ echo "$OLD_VERSION ->> $NEW_VERSION"
 if [ "b$NEW_VERSION" != "b $2$TWEAK" ]
 then
     echo "FAIL: Could not update from version $1 to version $2"
-    dropdb ___test_update
+    dropdb "${DB}"
     exit 1
 fi
 
@@ -114,10 +107,13 @@ DIR="sql/sigs"
 FILE="$DIR/pgrouting--$2.sig"
 
 echo "#VERSION pgrouting $2" > "$FILE"
-echo "#TYPES" >> $FILE
-psql $DB -c '\dx+ pgrouting' -A | grep '^type' | cut -d ' ' -f2- | sort >> $FILE
-echo "#FUNCTIONS" >> "$FILE"
-psql $DB -c '\dx+ pgrouting' -A | grep '^function' | cut -d ' ' -f2- | sort >> $FILE
+{
+    echo "#TYPES"
+    psql "${DB}" -c '\dx+ pgrouting' -A | grep '^type' | cut -d ' ' -f2-
+    echo "#FUNCTIONS"
+    psql ${DB} -c '\dx+ pgrouting' -A | grep '^function' | cut -d ' ' -f2-
+} >> "${FILE}"
+
 
 DIFF=$(git diff "$FILE")
 if [ -n "$DIFF" ]
@@ -138,7 +134,4 @@ dropdb "$DB"
 } # end of function
 
 
-update_test $FROM_PGR $CURRENT
-exit
-
-exit 0
+update_test "${FROM_PGR}" "${CURRENT}"