+++ tools/transifex/download_translation.sh
@@ -6,10 +6,6 @@
 # Download translations and build localized documentation
 # ------------------------------------------------------------------------------
 
-ROOT=$(pwd)
-CONFIG="build/doc/_build"
-DOCDIR="doc/i18n"
-
 LANGUAGES='de es ja fr'
 
 if [ "$1" ]; then